test(input): migrate item
