+++ core/src/components/input/test/item/input.e2e.ts-snapshots/input-inset-list-no-fill-ios-ltr-Mobile-Chrome-linux.png
