+++ core/src/components/input/test/item/input.e2e.ts-snapshots/input-list-no-fill-ios-rtl-Mobile-Chrome-linux.png
