+++ core/src/components/input/test/item/input.e2e.ts-snapshots/input-list-outline-ios-ltr-Mobile-Firefox-linux.png
