+++ core/src/components/input/test/item/input.e2e.ts-snapshots/input-list-solid-ios-ltr-Mobile-Chrome-linux.png
