+++ core/src/components/input/test/item/input.e2e.ts-snapshots/input-list-solid-ios-rtl-Mobile-Safari-linux.png
